pyproject.toml
[tool.poetry]
name = "canceled-orders"
version = "1.0.0"
description = ""
authors = ["Eldos <eldos.baktybekov@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "canceled_orders" }]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
bs4 = "^0.0.2"
lxml = "^5.2.1"
faststream = { extras = ["rabbit"], version = "^0.5.3" }
fast-depends = "^2.4.2"
structlog = "^24.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
factory-boy = "^3.3.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
    "src",
]
